requirements.txt
Django==4.2.6
djangorestframework==3.14.0
django-filter==23.3
django-countries==7.5.1
setuptools==68.2.2